No reward exploration debug
if there's no reward,
Alpha must be lower than .3
k is optimal around 250
alpha decrease changes results as well

+++ QLearning_FrameWork/Q_update.cpp
@@ -3,7 +3,7 @@
 #include "macros.h"
 
 extern float Q[5][5][8]; 				//Quality of taking an action state pair
-extern unsigned int N[5][5][8];			// Number of times you've taken an action state pair
+extern float N[5][5][8];			// Number of times you've taken an action state pair
 
 extern int Valid_Actions[8]; 		
 extern int Valid_Actions_Prime[8]; 
@@ -40,10 +40,22 @@ void Q_Update(){
 
   float oldPerspective;
   float newPerspective;
-
+//// debug prints---------------------------------------------------------------------------------------------------
+// Serial.print("Updating States: "); Serial.print(S1); Serial.print(", "); Serial.print(S2); Serial.println(" :");
+// Serial.print("Q(s,s:   "); Serial.print("a"); Serial.print(") -> "); 
+// Serial.print("beta"); Serial.print("*("); Serial.print("Q(a)"); Serial.print(") + ");
+// Serial.print("alpha"); Serial.print("*(("); Serial.print("R"); Serial.print(") + ("); Serial.print("qmax"); Serial.print(") + ((");
+// Serial.print("K"); Serial.print(" / "); Serial.print("N(a')"); Serial.print(")) = "); Serial.println("Q(s,s,a)updated;");
+//// debug prints^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ 
   for(int i=0; i<Actions_Num; i++){		//Update all possible actions' quaility 
   	Action_Being_Updated = Valid_Actions[i];  //pull from  global valid action vector
-
+   
+//// debug prints---------------------------------------------------------------------------------------------------
+//    Serial.print("Q(action "); Serial.print(Action_Being_Updated); Serial.print(") -> "); 
+//    Serial.print(beta); Serial.print("*("); Serial.print(Q[S1][S2][Action_Being_Updated]); Serial.print(") + ");
+//// debug prints^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+    
+    
   	S_Given_A(S1, S2, Action_Being_Updated);     //translate watched action into a state set
   	Available_Actions(S1_Prime,S2_Prime, 1); //Then update an array of actions in that state
 
@@ -51,12 +63,21 @@ void Q_Update(){
   	qmax = Q[S1_Prime][S2_Prime][A_Prime]; //store the value of that action
 
   	Reward = (Action_Being_Updated == Action_Next) ? R : 0; //if looking at the actual action taken, factor in Reward
+//// debug prints---------------------------------------------------------------------------------------------------
+//    Serial.print(alpha); Serial.print("*(("); Serial.print(R); Serial.print(") + ("); Serial.print(qmax); Serial.print(") + ((");
+//    Serial.print(K); Serial.print(" / "); Serial.print(N[S1_Prime][S2_Prime][A_Prime]); Serial.print(")) = ");
+//// debug prints^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
   	oldPerspective = Q[S1][S2][Action_Being_Updated];  //store current map value
   	newPerspective = Reward + qmax + K/N[S1_Prime][S2_Prime][A_Prime]; //store information learned
 
   	Q[S1][S2][Action_Being_Updated] = beta*oldPerspective + alpha*newPerspective; //Combined current map with learned info
   																				//beta and alpha weight old vs new
+//// debug prints---------------------------------------------------------------------------------------------------
+//    Serial.print(Q[S1][S2][Action_Being_Updated]); Serial.println(";");
+//// debug prints^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+                                  
   }
+//  Serial.println();
 }
